Expose IO.ic and IO.oc types for Cohttp_async.
Without this, it isn't possible to use Request.read_form, or anything where you
need to be able to construct an ic/oc. Cohttp.S.IO has ic and oc
abstract, and previously Cohttp_async only exposed the t type, meaning
that constructing a Reader.t / Writer.t (what the types actually are)
still fails to compile.

This change makes Cohttp_async use the same pattern as Cohttp_lwt_unix,
which is to give a signature to the concrete IO module that exposes ic,
oc, and conn (instead of just t), and use that in defining the
Request/Response signatures. A single test exercising this is
included.