Cohttp_lwt.Client: fix handling of HEAD request bodies, fix body non-consumption
read_response now requires a method and, if the method is `HEAD, no body
will be read. A method stream is produced from a request pipeline and this
result is now bound to the Lwt monad. A last_body reference has been introduced
which allows the response stream to finish drainage of any unhandled body.

Fixes #313. Fixes #315.