Re-organize some stuff
The re-organization is done for 2 reasons:

* To help with the eventual split of cohttp into multiple opam packages
* To follow jbuilder convention for the eventual port

To do this, we should prefer to have only 1 library per directory. And we should
name the opam files according to the packages our repository provides.

Also, this change moves the uri.services dependencies to cohttp_async. Which is
the only place where it's used.