Fix exception observed on 32-bit architectures
When determining how much to read in one go in a fixed length
transfer, we used to first convert an Int64 length to an int
and then bound that by 0x8000. However, the conversion can
result in a negative value and this later on results in an
exception. The 0x8000 bound should be applied in Int64 space.