Async client: do not read body if there is none
Responses with code 204 (No content) and 304 (Not modified) do not have
bodies. They can come with no Content-Length header, in which case the
transfer encoding is inferred as `Transfer.Unknown`. When trying to read
the nonexistent body, it would do so using `Transfer_io.Unknown.read`,
which attempts to read until EOF, which can cause hangs.

There is some logic in `Response.has_body` to determine if we should
attempt to read a body at all. This was hooked in the `Lwt` backend but
not in the `Async` one.

Closes #666