Don't use Gc.finalise for closing fd or consuming stream
It may run from any thread, which forbids calling into Lwt from it.

The best alternative I can currently think of is just issueing a warning
in case the stream is not closed yet when stream is collected.
Not even callin Log.warn is permitted, since it also uses Lwt.