cohttp.Header: modify update function to use Map.update
Thanks @pm-mck for the suggestion. The function [f] accepts a
[string option] even though some keys can be associated to lists.
The version I wrote only acceptin a [string list] was messy and
the one accepting [string list option] allows users to associate list of values to headers that should not accept them (note that you can still inject lists of values using map).

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>