cohttp-lwt: log the body leak
instead of printing it on stderr. Maybe the current way is better,
since it is explicit. But I think it makes more sense to have this
logged nicely instead of having random prints around.

This also extends the documentation a bit to explicitly mention the
need to drain the body.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>