cohttp-eio: convert to Eio.Buf_read
This saves a lot of code and avoids embedding another parser combinator
library in cohttp.

I tested it with `wrk -t 4 -c 1000 -d 5s http://127.0.0.1:8080` while
running the example server, and performance seems to be unaffected

I also changed the on_error handler to not say the error happened on
accept, since it's actually used for errors after accept.