Http.Headers: fix semantics of 'first'
The headers are dealt with in reverse order for speed and reversed
once converted back to list. This means that to move an element to
the top of the headers we need to attach it to the bottom of the
list.  The  private method has been mofied accordingly.

The bug was spotted when fixing the null-header test to address the
lack of version in the user-agent header for development versions
of cohttp

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>