cohttp-eio: add Client.make_generic and HTTPS support
The new `Client.make_generic` allows the user to provide their own
function for resolving URIs to flows.

The convenience wrapper `make` now taken an `https` argument,
which is a function that can be used to wrap raw sockets with a TLS
library.

The `client_tls.ml` example shows how to use this with tls-eio.