Expose the Socketaddr when constructing a socket
`cohttp-lwt-unix`'s caches based on `Conduit.endp` (aliased via its `Net.endp`).
The equivalent in Eio is `Eio.Net.Sockaddr.stream`.

We need to cache based on the Sockaddrs that are used here
https://github.com/tarides/ocaml-cohttp/blob/5465d00bfcee55fa077e0cd1f19964457bb88db6/cohttp-eio/src/client.ml#L53-L61
which means we need to a way of retrieving this bit of data when we
construct a socket.