Add a connection abstraction
We aim to cache connections. But what do we count as a connection?

When making requests, `Eio.Buf_write.with_flow socket f` opens the
`socket` for writing, and takes care of ensuring that the flow is closed
as a finalization after `f` completes. While we have the socket open for
writing, we also open the `socket` for reading. This is our
bidirectional connection, but it is only kept open as long as `f` runs.

To cache this open socket, but make it possible to reuse it for multiple
requests, we do the following:

- We make each connection a forked process that can read from a channel
  (an `Eio.Stream`).
  - Each message sent on the stream includes a resolver for a promise,
    which is how the response is returned.
- The process itself loops on reads from the channel, until the channel
  is closed.
- Since each processes is forked
  - it will get cleaned up when the enclosing switch is cancelled
  - but we can also close all the channels to tell them to shut down

This is all abstracted into the new `Connection` API (echoing the same
API in cohttp-lwt).

Our cache will then be a map `Socketaddr -> Connection.t`.